# `DensityOutliersTransform` optimization