--- a/etna/analysis/outliers/hist_outliers.py
+++ b/etna/analysis/outliers/hist_outliers.py
@@ -316,6 +316,8 @@ def get_anomalies_hist(
         name of the column in which the anomaly is searching
     bins_number:
         number of bins
+    index_only:
+        whether to return only outliers indices. If `False` will return outliers series
 
     Returns
     -------