--- a/etna/transforms/outliers/base.py
+++ b/etna/transforms/outliers/base.py
@@ -100,9 +100,11 @@ class OutliersTransform(ReversibleTransform, ABC):
         """
         if self.outliers_timestamps is None:
             raise ValueError("Transform is not fitted! Fit the Transform before calling transform method.")
-        segments = df.columns.get_level_values("segment").unique().tolist()
+        segments = set(df.columns.get_level_values("segment"))
         check_new_segments(transform_segments=segments, fit_segments=self._fit_segments)
         for segment in self.outliers_timestamps:
+            if segment not in segments:
+                continue
             # to locate only present indices
             segment_outliers_timestamps = df.index.intersection(self.outliers_timestamps[segment])
             df.loc[segment_outliers_timestamps, pd.IndexSlice[segment, self.in_column]] = np.NaN
@@ -131,9 +133,11 @@ class OutliersTransform(ReversibleTransform, ABC):
         """
         if self.original_values is None or self.outliers_timestamps is None:
             raise ValueError("Transform is not fitted! Fit the Transform before calling inverse_transform method.")
-        segments = df.columns.get_level_values("segment").unique().tolist()
+        segments = set(df.columns.get_level_values("segment"))
         check_new_segments(transform_segments=segments, fit_segments=self._fit_segments)
         for segment in self.outliers_timestamps:
+            if segment not in segments:
+                continue
             segment_ts = df[segment, self.in_column]
             segment_ts[segment_ts.index.isin(self.outliers_timestamps[segment])] = self.original_values[segment]
         return df