## Before submitting (must do checklist)

- [x] Did you read the [contribution guide](https://github.com/etna-team/etna/blob/master/CONTRIBUTING.md)?
- [x] Did you update the docs? We use Numpy format for all the methods and classes.
- [x] Did you write any new necessary tests?
- [ ] Did you update the [CHANGELOG](https://github.com/etna-team/etna/blob/master/CHANGELOG.md)?
<!-- For CHANGELOG separate each item in unreleased section by blank line to reduce collisions -->

## Flame Graphs (before)
![image](https://github.com/etna-team/etna/assets/60392282/0d1a6f4d-81e5-41f6-8d6d-c2acb0e50fbc)


## Proposed Changes
<!-- Add a more detailed description of the changes if needed. 
No need for typos and docs improvements  -->

- Partial vectorization for `get_segment_density_outliers_indices`
- Change data access in `get_anomalies_density`
- Store only segments with detected outliers
- Remove `_save_original_values` method

## Flame Graphs (after)
* Custom distance function `fit_transform`: 38.10s -> 29.89s
![image](https://github.com/etna-team/etna/assets/60392282/3822a81f-33ca-45db-a63e-aa55b09b2cd5)

* Provided distance function `fit_transform`: 38.10s -> 19.10s
![image](https://github.com/etna-team/etna/assets/60392282/98a44099-2c12-461d-9d4f-8d3d2f8d85d3)



## Additional thoughts

- Explore possibility of further vectorization
- Try to use compilation with numba (jit, ahead of time)

## Code to reproduce
```python
import pandas as pd
import numpy as np

from etna.datasets import TSDataset, generate_ar_df
from etna.transforms import DensityOutliersTransform


def add_anoms(df: pd.DataFrame, anom_frac: float):
    mask = np.random.choice([True, False], size=df.shape[0], p=[anom_frac, 1 - anom_frac])
    df.loc[mask, "target"] *= 1000
    return df


df = add_anoms(generate_ar_df(n_segments=10000, periods=100, start_time="2000-01-01"), anom_frac=0.01)

ts = TSDataset(df=TSDataset.to_dataset(df), freq="D")

transform = DensityOutliersTransform(
    in_column="target",
    window_size=15,
    distance_coef=3,
    n_neighbors=1,
    # distance_func=lambda x, y: abs(x - y)
)
transform.fit_transform(ts)
```

## Closing issues
<!-- Link Issue you are closing here. 
Put `closes #XXXX` in your comment to auto-close the issue that your PR fixes (if such). -->